v20.12.01 / sv1.7.0 - The first update since April!
All the changes will be explained in the changelog, but this is the first release since 15 April 2020, nearly 8 months ago. sv2.0.0 is currently being worked on, and that will include a major overhaul to both the UI *and* the backend itself, as we are making a move toward BrowserView. 

Enjoy!